Make functions directory relative.
Netlify uses

```
const source = path.resolve(process.cwd(), sourceArg)
```

to identify the source directory of the build functions. The previous 'functions' config value was not a relative path but a absolute path e.g. '/function' (absolute) to './function' (relative).

The last successful deploy was on [Mar 1 at 1:43 PM](https://app.netlify.com/sites/nuccobrain-staging/deploys/5c793708b583ef000893afc2) and the logs make no reference of 'Zipping functions from ...' which is referenced in the failing builds. Since that deploy [this](https://github.com/netlify/zip-it-and-ship-it) repo has been created which seems like a new tool netlify use internally to zip up the function artifacts and ship them off to aws lambda.

They did indeed just start using it and I can see that there is a difference in how they resolve the paths in both programs. netlif-lambda uses:

```
path.join(process.cwd(), functionsDir)
```

zip-it-and-ship-it uses

```
path.resolve(process.cwd(), sourceDir)
```

The difference is, the first simply concatenates them together and the second looks to see if the folder actually exists on the machine. This is a problem for us as it resolves paths properly (instead of just naively mashing them together) since our path is absolute (it starts with a slash). This means that `path.resolve` will start looking from the root of the machine.

This seems like a vulnerability actually, I will let netlify know.